ListBuilder, InventoryMasterList quickfix
It seems that on Mac, a GridColumn's DisplayIndex initializes correctly at instantiation, but on Windows, every GridColumn's DisplayIndex is -1 (even if there are multiple columns on a single grid) until some later time (possibly after the Shown event, but I didn't bother to confirm that). I was using a foreach loop to setting the header(s) for the LeftGridView in ListBuilder, relying on the DisplayIndex property; worked fine on Mac, ArrayBoundsException on Windows. It's now a regular for loop.

Like a dummy, I changed the filename for the IML but did not update the reference to it in the LoadInventoryMasterList() method. This has been fixed.
	modified:   LaSSI/ListBuilder.cs
	modified:   LaSSI/MainForm.eto.cs