Added Apply, Revert Buttons
Moved the entire data section (splitter and children) into its own class: DataPanel.
Created the class DetailsLayout to store the dynamic layout (and children) for each Details panel as well as a status (unmodified, modified, applied).
Added a layout for the Apply & Revert buttons to the Panel2 layout on the splitter (at the top, unfortunately).
Added some infrastructure for color-coding TreeGridItems.
Removed the Shift key from the Save-As command temporarily.

	new file:   DataPanel.cs
	new file:   DetailsLayout.cs
	modified:   MainForm.eto.cs