Finished apply/revert, DERELICTS
Added a ton of garbage related to tracking changes in the details panels. All of it needs to be revised; some needs to be rewritten or just ripped out.

Finished Apply; SaveAs now prompts the user to apply all unapplied changes.

Added the Tools menu with one tool so far: CLEAN DERELICTS.

	new file:   CollectionChange.cs
	modified:   DataPanel.cs
	modified:   DetailsLayout.cs
	modified:   InventoryGridItem.cs
	modified:   ListBuilder.cs
	modified:   MainForm.eto.cs
	new file:   Oncler.cs
	new file:   RadioInputDialog.cs
	new file:   TextInputDialog.cs