Issue #55
Suppressed a _lot_ of null warnings.

Added PromptForm class that is currently unused (and may never be).

Revised (to an extent) how the data status is tracked (added an enum called DataState).

Revised the prompt to apply changes before saving.

Added a prompt to save before quitting or opening another file (#55)

	modified:   CollectionChange.cs
	modified:   CustomCommands.cs
	modified:   DataPanel.cs
	modified:   ListBuilder.cs
	modified:   MainForm.cs
	modified:   MainForm.eto.cs
	new file:   PromptForm.cs