Mac file perms/attr step (Issue #64)
Finally gave up on trying to get the Mac file perms/attr stuff done in the workflow. The xattr step can't be done until _after_ the file is downloaded to a Mac client anyway, and the execute perms can't be done on Windows and I just don't feel like continuing to fuck around trying to get multiple runners working.

Also slightly revised the wording on the About screen.

	modified:   LaSSI/AboutCommand.cs
	new file:   LaSSI/MacBits.sh