Issues #4, #92, #93
Issue #94: moved the Closable=false to the Mac part of the PlatformSpecificNonsense() method

Issue #92: the Prefs menu item is now added to the File menu on Windows

Issue #4: implemented Preferences, primarily through the Prefs class. It has some flaws, but it will load default prefs (currently the only ones supported are always/never/prompt behavior for apply before save and save before close and start-up behavior (nothing, open file chooser, reopen last file, and load file)), check for a file in the user’s local app support directory, load any changes from there, respect the specified behaviors when starting and quitting the app and applying and saving changes, and save the current Pref values to the file. It’ll be extensible to a point, but it could definitely use some love and attention down the line.

Minor tweaks cleanup to some other classes.

	modified:   CustomCommands.cs
	modified:   DataPanel.cs
	new file:   FileMagnifyingGlassIcon24x24.ico
	new file:   FileMagnifyingGlassIcon24x24.png
	modified:   MainForm.eto.cs
	modified:   Prefs.cs
	modified:   RadioInputDialog.cs
	modified:   SaveFile.cs