Issue #69
Each time a file is opened, a copy (prepended with "orig") is moved to a timestamped backup directory in the app support directory. Each time the file is saved, a copy of the saved file is moved into the same directory (with an incrementing prepended int). If a new file is loaded or the app is quit without additional saves being created and moved into the backup directory, it is removed.

The app support directory can be easily access by a new File menu item called "Browse backups".

	modified:   CustomCommands.cs
	modified:   MainForm.eto.cs
	modified:   Modal.cs
	modified:   Prefs.cs