Issues #100, #102, #101, #97, #96
Removed the startup option "ShowFileChooser" since it causes a crash on Windows (and causes a weird window positioning issue on Mac).

Set a Win32 file icon.

Removed erroneous addition of second Settings menu item on Windows.

Overhauled CometExists() and ResetComet() to deal with the fact that the comet might be in a system that isn't archived (either because it's where we are or it hasn't been visited yet).

Fixed up the backup creation/addition/cleanup.

	modified:   CustomCommands.cs
	modified:   DataPanel.cs
	modified:   LaSSI.csproj
	modified:   MainForm.eto.cs
	modified:   Prefs.cs