Issue #43
Dec 20-31, the tree turns red and green; Jan 1-3, the tree turns brown; then it’s back to black. This can be turned off in the Prefs menu.

Apparently now we don’t need to get ready for quit as a separate step on Mac either. I don’t know why.

	modified:   CustomCommands.cs
	modified:   DataPanel.cs
	modified:   MainForm.eto.cs
	modified:   Prefs.cs