Removed explicit text color setting (in place of PR #128)
Removed unnecessary explicit setting of the tree view text to black—it caused minor problems for lightmode users (the selected node was hard to read) and major problems for darkmode users (everything was hard to read!) and didn't even fix the problem it was meant to (which may never have actually existed in the first place because I can't repro it now).

The other problem raised in PR #128 (the New Year holiday color being almost impossible to read in darkmode) will be addressed through workitem #48.

	modified:   DataPanel.cs