Issue #133
Backups are now stored in the backups subdir of the LaSSI app support
directory.

Also, any old backup directories are automatically moved into that
subdir on app close. This probably ought to be dropped in a while after
enough time has passed that no one still has backups in the old
location.

	modified:   CustomCommands.cs
	modified:   Prefs.cs