Issues #136, #108, #53
The Node class now has a baseName field and a method RebuildName(); need to do a survey to see if the method needs to be called anywhere currently (and be followed by a tree rebuild).

Rescue missions now have the correct detail notes "contract available: System x" and "ship location: System y".

Add a tool to change the Type of any "Stranded Ship" not associated with a Rescue mission or associated with a completed Rescue mission to "Derelict" so they'll get cleaned up, either by the game or by the Clean Derelicts tool.

	modified:   CustomCommands.cs
	modified:   DataPanel.cs
	modified:   Node.cs