Issue #135
Added basic tool to reassign missions: a check box list dialog to select one or more missions currently already assigned to ships in the fleet followed by a radio button dialog to select a ship to assign the missions to. Not necessarily what I'd prefer, but it'll do.

Also fixed a bug in the AssertionFailureConditionExists() method. Not sure when that was added or if it simply never worked.

RadioInputDialog now starts was the SelectedIndex as -1 and the OK button starts disabled (enabled once a selection is made).

CheckBoxListDialog new adds two buttons (All and None) above the list if there are more than 3 items in the list.

Minor revisions to the LassiReport class.

	modified:   CheckBoxListDialog.cs
	modified:   CustomCommands.cs
	modified:   DataPanel.cs
	modified:   LassiReport.cs
	modified:   RadioInputDialog.cs