WIP - check for updates
The Modal class is now set to be Topmost (to avoid situations where the modal could be lost behind the main window which is locked by the modal requiring the user to force-quit the application). This is not idea since now the modal window will float above *every* window on the system rather than just the main window.

The Modal class also now has a ModalShown() event method to set the location centered in the parent window correctly--requires that the Parent be set ahead of time *from the parent context*.

Revised verbiage in MainForm.CheckForUpdatesAsync() for new version/no new version alerts.

	modified:   MainForm.eto.cs
	modified:   Modal.cs