WIP on #186
The core is in place but diabled since updating a file or adding a file to the Steam cloud requires modifying a manifest file (including listing a SHA), so I'm tabling the feature for now as out of scope for the already-overdue next release.

Also minor update to VS styles.