Fixed #182
Datagrid context menu items enabled/disabled if a row is selected or not.
Reverted padding change to the scrollable on CheckBoxListDialog.
Fixed bug where canceling the "specify new key" prompt on a key collision when adding a new row to the datagrid caused a new row with a blank key to be added anyway.